// src/interfaces/meeting_rooms.interface.ts
export interface MeetingRooms {
    room_id: string;
    room_address: string;
    staff_id: string;
    date_created: string;
    state: string;
    lga: string;
    ward:string;
    hub_id: string;
    room_image: string;
    longitude: string;
    lattitude: string;
    delete_flag: number;
    type: string;
    imei: string;
    app_version: string;
  }
  